update Container & Container type, add Damage Type

+++ src/db/models/damage_type.js
@@ -20,7 +20,7 @@ module.exports = (sequelize, DataTypes) => {
 		{
 			dycode: DataTypes.STRING,
             dydesc: DataTypes.STRING,
-            dyclean: DataTypes.TINYINT,
+            dyclean: DataTypes.INTEGER,
 		},
 		{
 			sequelize,
